+++ doc/changelog.md
@@ -1,3 +1,10 @@
+# 0.2
+ - Started new theme fallen
+ - Added new example page
+ - Added config option createInnerTile : false to tilewin.js
+ - Added config option transition : "all 0.2s ease-in-out", to tilewin.js
+ - Bugfixed element.js
+ - Renamed the style element made my style.js to RUITKStyles
 ## 0.1.7
  - Updated readme
  - Renamed Project to RUITK